Loading in leapseconds kernel in Metakernel API (#1172)
* Loading in leapseconds before converting YYYYMMDD to J2000
