+++ tests/lambda_endpoints/test_spice_metakernel_api.py
@@ -1,10 +1,10 @@
 """Tests for the SPICE Query API."""
 
 import json
+import os
 from datetime import datetime, timedelta
 
 import imap_data_access
-import pytest
 
 from sds_data_manager.lambda_code.SDSCode.api_lambdas import spice_metakernel_api
 from sds_data_manager.lambda_code.SDSCode.database import models
@@ -275,10 +275,23 @@ def test_metakernel_filtered_file_types(session):
     assert result["body"] == "No files found."
 
 
-@pytest.mark.skip(reason="Need to fix the spiceypy.datetime2et() call")
-def test_metakernel_string_input(session):
+def test_metakernel_string_input(session, s3_client):
     """Test that string input is allowed, and is converted to a datetime object."""
-    _insert_test_file(session, "naif0012.tls", [[1, 300]], upload_time=1)
+    current_path = os.path.dirname(os.path.abspath(__file__))
+    one_level_up = os.path.abspath(os.path.join(current_path, ".."))
+    test_spice_data_dir = os.path.join(one_level_up, "test-data", "test_spice_files")
+
+    # Insert leapsecond spice kernel into a mock S3 bucket
+    lsk_test_path = os.path.join(test_spice_data_dir, "naif0012.tls")
+    bucket_name = os.getenv("S3_BUCKET")
+    with open(lsk_test_path, "rb") as f:
+        s3_client.put_object(
+            Bucket=bucket_name,
+            Key="imap/spice/lsk/naif0012.tls",
+            Body=f,
+        )
+
+    _insert_test_file(session, "imap/spice/lsk/naif0012.tls", [[1, 300]], upload_time=1)
     _insert_test_file(session, "imap_sclk_0012.tsc", [[1, 300]], upload_time=1)
     _insert_test_data(session)
 
@@ -293,7 +306,7 @@ def test_metakernel_string_input(session):
         },
         None,
     )
-    assert len(json.loads(result["body"])) == 8
+    assert len(json.loads(result["body"])) == 6
 
 
 def test_metakernel_frames(session):
